Fixed wrong datatype for stock price

+++ Data/SubjectData.cs
@@ -3,11 +3,11 @@ namespace StockMarketApp.Data
 {
     public class SubjectData
     {
-        public int Price { get; set; }
+        public double Price { get; set; }
         public string Name { get; set; }
         public string Ticker { get; set; }
 
-        public SubjectData(int price, string name, string ticker)
+        public SubjectData(double price, string name, string ticker)
         { 
             Price = price;
             Name = name;
